Refactor XGBoost JSON to accommodate additional SAX parser
